Fix media/base 64-bit compilation issues
BUG=166496
TBR=scherkus@chromium.org
Review URL: https://codereview.chromium.org/12043004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@177924 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a56a718970c25ce59ec2eece48b3309bef260b0e

+++ media/base/audio_bus.h
@@ -83,7 +83,7 @@ class MEDIA_EXPORT AudioBus {
   const float* channel(int channel) const { return channel_data_[channel]; }
   void SetChannelData(int channel, float* data);
 
-  int channels() const { return channel_data_.size(); }
+  int channels() const { return static_cast<int>(channel_data_.size()); }
   int frames() const { return frames_; }
   void set_frames(int frames);
